Update baselines and conditionalize tests
